% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Cytoscape.R
\name{AdjMatrix2CytoNodes}
\alias{AdjMatrix2CytoNodes}
\title{Push Adjacency Matrix to Cytoscape}
\usage{
AdjMatrix2CytoNodes(Adj, compute = "dina", title = "Scenario3_AdjMatrix")
}
\arguments{
\item{Adj}{A data.frame Adjacency Matrix setup.}

\item{compute}{Character. Default compute rule.}

\item{title}{Character. The Network Title in Cytoscape.}
}
\value{
SUID of pushed Cytoscape Network.
}
\description{
Push Adjacency Matrix to Cytoscape
}
