% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Cytoscape.R
\name{AdjMatrix2iGraph}
\alias{AdjMatrix2iGraph}
\title{Convert Adjacency Matrix directly to igraph}
\usage{
AdjMatrix2iGraph(Adj, compute = "dina")
}
\arguments{
\item{Adj}{A data.frame adjacency matrix evaluating structural transitions.}

\item{compute}{Character. Default compute rule applied iteratively.}
}
\value{
A directed \code{igraph} object.
}
\description{
Convert Adjacency Matrix directly to igraph
}
