% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Constructors.R
\name{build_from_adjacency}
\alias{build_from_adjacency}
\title{Build Graph from an Adjacency Matrix}
\usage{
build_from_adjacency(
  AdjMatrixFile,
  DataFileCols,
  DefaultAttributeCompute = "zscore",
  DefaultTaskCompute = "dina"
)
}
\arguments{
\item{AdjMatrixFile}{Character. Path to the CSV file containing the square adjacency matrix.}

\item{DataFileCols}{Character vector. Names of columns present in the observational data, used to map "Task" nodes.}

\item{DefaultAttributeCompute}{Character. Default compute rule for attributes. Default is "zscore".}

\item{DefaultTaskCompute}{Character. Default compute rule for tasks. Default is "dina".}
}
\value{
A topologically sorted \code{igraph} object.
}
\description{
Constructs a directed \code{igraph} object from an adjacency matrix.
Identifies "Task" nodes based on columns present in the dataset.
}
