% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Constructors.R
\name{build_from_node_edge_files}
\alias{build_from_node_edge_files}
\title{Build Graph from Node and Edge Files}
\usage{
build_from_node_edge_files(NodesFile, EdgesFile)
}
\arguments{
\item{NodesFile}{Character. Path to the CSV file containing node definitions.}

\item{EdgesFile}{Character. Path to the CSV file containing edge definitions.}
}
\value{
A topologically sorted \code{igraph} object.
}
\description{
Constructs a directed \code{igraph} object from separate node and edge CSV lists.
}
