% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Constructors.R
\name{build_from_q_matrix}
\alias{build_from_q_matrix}
\title{Build Graph from a Q-Matrix}
\usage{
build_from_q_matrix(
  QMatrixFile,
  DefaultAttributeCompute = "zscore",
  DefaultTaskCompute = "dina"
)
}
\arguments{
\item{QMatrixFile}{Character. Path to the CSV file containing the Q-matrix.}

\item{DefaultAttributeCompute}{Character. Default compute rule for attributes. Default is "zscore".}

\item{DefaultTaskCompute}{Character. Default compute rule for tasks. Default is "dina".}
}
\value{
A topologically sorted \code{igraph} object representing the Q-matrix structure.
}
\description{
Constructs a bipartite \code{igraph} object from a standard Q-matrix file.
}
