% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Constructors.R
\name{build_irt_graph}
\alias{build_irt_graph}
\title{Build Graph for an IRT Model}
\usage{
build_irt_graph(
  task_names,
  ability_name = "Theta",
  default_task_compute = "dina"
)
}
\arguments{
\item{task_names}{Character vector. Names of the tasks (columns from the dataset).}

\item{ability_name}{Character. Name of the single latent trait. Default is "Theta".}

\item{default_task_compute}{Character. Default compute rule for tasks. Default is "dina".}
}
\value{
A topologically sorted \code{igraph} object representing the IRT structure.
}
\description{
Constructs a single-attribute \code{igraph} object directly from a list of task names,
bypassing the need for a Q-matrix file. All tasks are assumed to require the single ability.
}
