% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config_Main.R
\name{build_scoring_config}
\alias{build_scoring_config}
\title{Build Model Configuration for Scoring}
\usage{
build_scoring_config(calib_results, calib_config, new_dataframe)
}
\arguments{
\item{calib_results}{The list returned from \code{run_pgdcm_auto} containing \code{mapped_parameters}.}

\item{calib_config}{The model configuration object used to generate the calibrated model.}

\item{new_dataframe}{A \code{data.frame} of raw responses matching the new participants to score.}
}
\value{
A configuration list prepared for \code{run_pgdcm_auto} execution.
}
\description{
Automatically isolates and structures the compiled structural and item constraints 
from a calibrated model into a tightly locked scoring model configuration.
}
