% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inferences_DiBello.R
\name{calc_risk_difference}
\alias{calc_risk_difference}
\title{Get Prerequisite Gate Strength (Risk Difference)}
\usage{
calc_risk_difference(results, skill)
}
\arguments{
\item{results}{The full pgdcm results list object.}

\item{skill}{A string name or integer index.}
}
\value{
A list containing the target name and posterior means/intervals for the gate strength.
}
\description{
Calculates the absolute difference between the Bottleneck Probability and the Leap Probability.
}
\section{Semantic Interpretation}{

* **High Value (> 0.6)**: The prerequisite is critically required. Mastering the prerequisite creates a massive jump in the likelihood of mastering the target skill.
* **Low Value (< 0.2)**: A weak structural connection. Having the prerequisite barely improves a student's chances of getting the target skill, implying the theoretical arrow in your DAG represents a weak or non-existent causal relationship.
}

