% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Statistics.R
\name{check_mcmc_convergence}
\alias{check_mcmc_convergence}
\title{Check MCMC Convergence}
\usage{
check_mcmc_convergence(chainlist, blocksize = 10, burninperiod = 1000)
}
\arguments{
\item{chainlist}{An \code{mcmc.list}, matrix, or a list containing samples.}

\item{blocksize}{Numeric. Size of the blocks to average over. Default is 10.}

\item{burninperiod}{Numeric. Number of initial samples to discard as burn-in. Default is 1000.}
}
\value{
A list containing \code{avgparamvector}, \code{abserrors}, \code{relerrors}, and a boolean \code{converged} flag.
}
\description{
Evaluates rough convergence metrics by comparing block averages.
}
