% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Editor.R
\name{checklegaledges}
\alias{checklegaledges}
\title{Check Legal Edges}
\usage{
checklegaledges(igraphnodes, igraphedges)
}
\arguments{
\item{igraphnodes}{A dataframe of nodes}

\item{igraphedges}{A dataframe of edges}
}
\value{
A list of bad edges
}
\description{
Check Legal Edges
}
