% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Statistics.R
\name{compute_M2}
\alias{compute_M2}
\title{Compute Second-Order Moment Matrix}
\usage{
compute_M2(mat)
}
\arguments{
\item{mat}{A numeric matrix of responses.}
}
\value{
A numeric matrix representing the second-order moments.
}
\description{
Computes the second-order moment matrix (uncentered covariance/co-occurrence)
for a given matrix (usually binary responses).
}
