% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config_SEM.R
\name{configure_sem}
\alias{configure_sem}
\title{Configure SEM Initialization and Constants}
\usage{
configure_sem(info, X)
}
\arguments{
\item{info}{Graph structural properties from \code{get_graph_info}.}

\item{X}{A numeric matrix representing the observational participant data.}
}
\value{
A list with \code{constants}, \code{inits}, \code{monitors}, and \code{data}.
}
\description{
Generates the necessary Nimble \code{constants}, \code{inits}, and \code{monitors}
tailored for Structural Equation Models (SEMs).
}
