% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Helpers.R
\name{create_edge_df}
\alias{create_edge_df}
\title{Create a Standardized Edge DataFrame}
\usage{
create_edge_df(from, to)
}
\arguments{
\item{from}{Character vector of source node identifiers.}

\item{to}{Character vector of target node identifiers.}
}
\value{
A \code{data.frame} formatted for graph construction.
}
\description{
Helper function to generate a well-formatted edge \code{data.frame}.
}
