% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Helpers.R
\name{create_node_df}
\alias{create_node_df}
\title{Create a Standardized Node DataFrame}
\usage{
create_node_df(name, type, compute = "dina")
}
\arguments{
\item{name}{Character vector of node identifiers.}

\item{type}{Character vector of node types (e.g., "Task", "Attribute").}

\item{compute}{Character vector indicating the compute rule (default is "dina").}
}
\value{
A \code{data.frame} formatted for graph construction.
}
\description{
Helper function to generate a well-formatted node \code{data.frame}.
}
