% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config_Parsers.R
\name{determine_model_type}
\alias{determine_model_type}
\title{Determine Model Type}
\usage{
determine_model_type(info)
}
\arguments{
\item{info}{A list of graph properties returned by \code{get_graph_info}.}
}
\value{
A character string denoting the model type ("SEM" or "DCM").
}
\description{
Classifies the type of model (SEM vs DCM) explicitly by evaluating
the structure of the latent attributes and tasks.
}
