% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Editor.R
\name{edgeadd}
\alias{edgeadd}
\title{Add Edge}
\usage{
edgeadd(theigraph, fromnodechoice, tonodechoice, colornickname)
}
\arguments{
\item{theigraph}{The igraph object}

\item{fromnodechoice}{The source node}

\item{tonodechoice}{The target node}

\item{colornickname}{The color nickname}
}
\value{
The updated igraph object
}
\description{
Add Edge
}
