% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Editor.R
\name{edgedelete}
\alias{edgedelete}
\title{Delete Edge}
\usage{
edgedelete(theigraph, fromnodechoice, tonodechoice)
}
\arguments{
\item{theigraph}{The igraph object}

\item{fromnodechoice}{The source node}

\item{tonodechoice}{The target node}
}
\value{
The updated igraph object
}
\description{
Delete Edge
}
