% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Helpers.R
\name{enforce_topo_sort}
\alias{enforce_topo_sort}
\title{Enforce Topological Sort on Attributes}
\usage{
enforce_topo_sort(g)
}
\arguments{
\item{g}{An \code{igraph} object containing "Attribute" and "Task" nodes.}
}
\value{
A new \code{igraph} object with vertices ordered topologically.
}
\description{
Topologically sorts the Attribute nodes in a graph and appends the Task nodes.
Useful for ensuring attributes are processed in dependency order.
}
