% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Cytoscape.R
\name{get_Cyto_template}
\alias{get_Cyto_template}
\title{Get Cytoscape Template}
\usage{
get_Cyto_template(dest_dir = getwd(), overwrite = FALSE)
}
\arguments{
\item{dest_dir}{Character. The destination directory to copy the template to. Defaults to the current working directory.}

\item{overwrite}{Logical. Whether to overwrite the file if it already exists. Defaults to FALSE.}
}
\value{
Logical. TRUE if successful, FALSE otherwise.
}
\description{
Copies the bundled PGDCM Cytoscape template file to the current working directory.
}
