% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config_Parsers.R
\name{get_graph_info}
\alias{get_graph_info}
\title{Extract Graph Information}
\usage{
get_graph_info(g)
}
\arguments{
\item{g}{An \code{igraph} object containing "Attribute" and "Task" nodes.}
}
\value{
A list containing properties such as \code{nrnodes}, \code{nrattributenodes},
  \code{nrtasknodes}, and boolean flags for compute rules (e.g., \code{isDINA}).
}
\description{
Evaluates the \code{igraph} object to extract dimensional structural properties,
adjacency matrices, and compute rules required for Nimble configuration.
}
