% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Cytoscape.R
\name{get_NodesTable}
\alias{get_NodesTable}
\title{Extract Nodes Table from Graph}
\usage{
get_NodesTable(graph)
}
\arguments{
\item{graph}{An \code{igraph} object.}
}
\value{
A \code{data.frame} containing the nodes with columns \code{id}, \code{type}, and \code{compute}.
}
\description{
Extracts the vertex attributes from an \code{igraph} object and formats them
into a standardized nodes \code{data.frame}.
}
