% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Editor.R
\name{LayoutChange}
\alias{LayoutChange}
\title{Change Layout}
\usage{
LayoutChange(theigraph, layoutchoicestring)
}
\arguments{
\item{theigraph}{The igraph object}

\item{layoutchoicestring}{The layout choice string}
}
\description{
Change Layout
}
