% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Statistics.R
\name{map_sem_parameters}
\alias{map_sem_parameters}
\title{Map SEM Parameters to Readable Names}
\usage{
map_sem_parameters(summary_mx, config_obj, student_names = NULL)
}
\description{
Map SEM Parameters to Readable Names
}
