% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Cytoscape.R
\name{QMatrix2CytoNodes}
\alias{QMatrix2CytoNodes}
\title{Push Q-Matrix to Cytoscape}
\usage{
QMatrix2CytoNodes(Q, compute = "dina", title = "Scenario2_QMatrix")
}
\arguments{
\item{Q}{A data.frame Q-Matrix setup.}

\item{compute}{Character. Default compute rule.}

\item{title}{Character. The Network Title in Cytoscape.}
}
\value{
SUID of pushed Cytoscape Network.
}
\description{
Push Q-Matrix to Cytoscape
}
