% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Cytoscape.R
\name{QMatrix2iGraph}
\alias{QMatrix2iGraph}
\title{Convert Q-Matrix directly to igraph}
\usage{
QMatrix2iGraph(Q, compute = "dina")
}
\arguments{
\item{Q}{A data.frame acting as the Q-Matrix, where the first column contains Task IDs and remaining columns are Attributes.}

\item{compute}{Character. Compute rule applied uniformly to tasks and attributes. Default is "dina".}
}
\value{
A directed \code{igraph} object.
}
\description{
Convert Q-Matrix directly to igraph
}
