% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Editor.R
\name{readigraph}
\alias{readigraph}
\title{Read Igraph from Files}
\usage{
readigraph(FilePrefix)
}
\arguments{
\item{FilePrefix}{The prefix for the nodes and edges files}
}
\value{
An igraph object
}
\description{
Read Igraph from Files
}
