% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inferences_DiBello.R
\name{resolve_item}
\alias{resolve_item}
\title{Resolve Human-Readable Item Name to Integer Index safely via Mapped Parameters}
\usage{
resolve_item(results, item)
}
\description{
Resolve Human-Readable Item Name to Integer Index safely via Mapped Parameters
}
