% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inferences_DiBello.R
\name{resolve_skill}
\alias{resolve_skill}
\title{Resolve Human-Readable Skill Name to Integer Index safely via Mapped Parameters}
\usage{
resolve_skill(results, skill)
}
\description{
Resolve Human-Readable Skill Name to Integer Index safely via Mapped Parameters
}
