% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Cytoscape.R
\name{Tasks2CytoNodes}
\alias{Tasks2CytoNodes}
\title{Push Standalone Task Nodes to Cytoscape}
\usage{
Tasks2CytoNodes(task_names, compute = "dina")
}
\arguments{
\item{task_names}{Character vector of task node IDs.}

\item{compute}{Character. Default compute rule. Default is "dina".}
}
\description{
Pushes an orphaned list of task nodes directly to Cytoscape without edges.
}
