% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config_Main.R
\name{validate_graph_compute_nodes}
\alias{validate_graph_compute_nodes}
\title{Validate Graph Compute Nodes}
\usage{
validate_graph_compute_nodes(graph)
}
\arguments{
\item{graph}{An \code{igraph} object.}
}
\value{
Boolean indicating if validation passed. Throws an error if unsupported nodes are found.
}
\description{
Ensures that all nodes in the graph have a supported compute property.
}
