% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Editor.R
\name{viewparentsandchildren}
\alias{viewparentsandchildren}
\title{View Parents and Children}
\usage{
viewparentsandchildren(theigraph, nodechoice)
}
\arguments{
\item{theigraph}{The igraph object}

\item{nodechoice}{The node to view}
}
\description{
View Parents and Children
}
