% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Editor.R
\name{writeigraph}
\alias{writeigraph}
\title{Write Igraph to Files}
\usage{
writeigraph(FilePrefix, theigraph)
}
\arguments{
\item{FilePrefix}{The prefix for the nodes and edges files}

\item{theigraph}{The igraph object to write}
}
\description{
Write Igraph to Files
}
