% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Cytoscape.R
\name{push_to_cytoscape}
\alias{push_to_cytoscape}
\title{Push Graph to Cytoscape}
\usage{
push_to_cytoscape(..., base.url = "http://localhost:1234/v1")
}
\arguments{
\item{...}{Arguments passed to \code{createNetworkFromDataFrames}.}

\item{base.url}{Character. The base URL for the Cytoscape REST API. Default is \code{"http://localhost:1234/v1"}.}
}
\value{
Invisible SUID of the created network in Cytoscape.
}
\description{
Sends the graph structure to a running Cytoscape instance for visualization.
}
